backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-binance>=1.0.19
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
